move Extension & Configuration interfaces impl to HangmanExtConf from Kernel
This is due to a twice Kernel::process launch when kernel is
autoregistered as extension

branch:'feature/hangman-event-dispatcher'
inherit:'master'
